!fix float create

+++ database/migrations/2023_09_14_063340_create_flats_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('longitude');
             $table->string('latitude');
             $table->foreignId('currency_id')->references('id')->on('currencies')->onDelete('cascade');
-            $table->float('price');
+            $table->decimal('price', 10);
             $table->float('price_per_meter')->nullable();
             $table->float('price_day')->nullable();
             $table->float('price_week')->nullable();
